consolidate insert edges
