--- src/agdb/query_builder/insert_edge_values.rs
@@ -1,9 +1,0 @@
-use crate::query::insert_edges_query::InsertEdgesQuery;
-
-pub struct InsertEdgeValues(pub InsertEdgesQuery);
-
-impl InsertEdgeValues {
-    pub fn query(self) -> InsertEdgesQuery {
-        self.0
-    }
-}
