--- src/agdb/query_builder/insert_edges_values.rs
@@ -1,9 +1,0 @@
-use crate::query::insert_edges_query::InsertEdgesQuery;
-
-pub struct InsertEdgesValues(pub InsertEdgesQuery);
-
-impl InsertEdgesValues {
-    pub fn query(self) -> InsertEdgesQuery {
-        self.0
-    }
-}
